+++ apps/mobile/src/components/episode/OfflineEpisode.tsx
@@ -131,7 +131,6 @@ export function OfflineEpisode({
             style={{
               height: 20,
             }}>
-            {/** @ts-expect-error wrong types **/}
             <GlueIcon as={TrashIcon} style={{ color: colors.error.color }} />
           </ActionsheetIcon>
         </ActionSheetItem>
