+++ apps/mobile/src/components/episode/index.ts
@@ -1,3 +1,4 @@
 export * from './Episode';
 export * from './QualitySelect';
 export * from './OfflineEpisode';
+export * from './EpisodeNumber.tsx';
